[beatreceiver] Add e2e test for journald, unix and winlog input
