--- x-pack/filebeat/tests/integration/otel_gcppubsub_test.go
@@ -1,174 +1,0 @@
-// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
-// or more contributor license agreements. Licensed under the Elastic License;
-// you may not use this file except in compliance with the Elastic License.
-
-//go:build integration
-
-package integration
-
-import (
-	"bytes"
-	"context"
-	"fmt"
-	"testing"
-	"text/template"
-	"time"
-
-	"github.com/gofrs/uuid/v5"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"github.com/elastic/beats/v7/libbeat/tests/integration"
-	"github.com/elastic/beats/v7/x-pack/filebeat/input/gcppubsub/testutil"
-	"github.com/elastic/beats/v7/x-pack/otel/oteltest"
-	"github.com/elastic/beats/v7/x-pack/otel/oteltestcol"
-
-	"github.com/elastic/elastic-agent-libs/testing/estools"
-)
-
-func TestGCPInputOTelE2E(t *testing.T) {
-	integration.EnsureESIsRunning(t)
-
-	// Create pubsub client for setting up and communicating to emulator.
-	client, clientCancel := testutil.TestSetup(t)
-	defer func() {
-		clientCancel()
-		client.Close()
-	}()
-
-	testutil.CreateTopic(t, client)
-	testutil.CreateSubscription(t, "test-subscription-otel", client)
-	testutil.CreateSubscription(t, "test-subscription-fb", client)
-	const numMsgs = 10
-	testutil.PublishMessages(t, client, numMsgs)
-
-	host := integration.GetESURL(t, "http")
-	user := host.User.Username()
-	password, _ := host.User.Password()
-
-	// create a random uuid and make sure it doesn't contain dashes/
-	otelNamespace := fmt.Sprintf("%x", uuid.Must(uuid.NewV4()))
-	fbNameSpace := fmt.Sprintf("%x", uuid.Must(uuid.NewV4()))
-
-	otelIndex := "logs-integration-" + otelNamespace
-	fbIndex := "logs-integration-" + fbNameSpace
-
-	type options struct {
-		Index           string
-		ESURL           string
-		Username        string
-		Password        string
-		Subscription    string
-		CredentialsFile string
-	}
-
-	gcpFilebeatConfig := `filebeat.inputs:
-- type: gcp-pubsub
-  project_id: test-project-id
-  topic: test-topic-foo
-  subscription.name:  {{ .Subscription }}
-  credentials_file: "{{ .CredentialsFile }}"
-` + filebeatOutputYAML
-
-	gcpOTelConfig := otelElasticsearchExporterYAML + `receivers:
-    filebeatreceiver:
-        filebeat:
-            inputs:
-                - credentials_file: "{{ .CredentialsFile }}"
-                  project_id: test-project-id
-                  subscription:
-                    name: {{ .Subscription }}
-                  topic: test-topic-foo
-                  type: gcp-pubsub
-        processors:
-            - add_host_metadata: ~
-            - add_cloud_metadata: ~
-            - add_docker_metadata: ~
-            - add_kubernetes_metadata: ~
-        queue.mem.flush.timeout: 0s
-        setup.template.enabled: false
-        management.otel.enabled: true
-` + otelElasticsearchServiceYAML
-
-	optionsValue := options{
-		ESURL:           fmt.Sprintf("%s://%s", host.Scheme, host.Host),
-		Username:        user,
-		Password:        password,
-		CredentialsFile: "testdata/gcp_pubsub_fake_credentials.json", //nolint:gosec // this is test file
-	}
-
-	var configBuffer bytes.Buffer
-	optionsValue.Index = otelIndex
-	optionsValue.Subscription = "test-subscription-otel"
-	require.NoError(t, template.Must(template.New("config").Parse(gcpOTelConfig)).Execute(&configBuffer, optionsValue))
-
-	oteltestcol.New(t, configBuffer.String())
-
-	// reset buffer
-	configBuffer.Reset()
-
-	optionsValue.Index = fbIndex
-	optionsValue.Subscription = "test-subscription-fb"
-	require.NoError(t, template.Must(template.New("config").Parse(gcpFilebeatConfig)).Execute(&configBuffer, optionsValue))
-
-	// start filebeat
-	filebeat := integration.NewBeat(
-		t,
-		"filebeat",
-		"../../filebeat.test",
-	)
-
-	filebeat.WriteConfigFile(configBuffer.String())
-	filebeat.Start()
-	defer filebeat.Stop()
-
-	// prepare to query ES
-	es := integration.GetESClient(t, "http")
-
-	t.Cleanup(func() {
-		deleteDataStreamsFromES(t, es, []string{otelIndex, fbIndex})
-	})
-
-	rawQuery := map[string]any{
-		"query": map[string]any{
-			"match_phrase": map[string]any{
-				"input.type": "gcp-pubsub",
-			},
-		},
-		"sort": []map[string]any{
-			{"@timestamp": map[string]any{"order": "asc"}},
-		},
-	}
-
-	var filebeatDocs estools.Documents
-	var otelDocs estools.Documents
-	var err error
-
-	// wait for logs to be published
-	require.EventuallyWithTf(t,
-		func(ct *assert.CollectT) {
-			findCtx, findCancel := context.WithTimeout(context.Background(), 10*time.Second)
-			defer findCancel()
-
-			otelDocs, err = estools.PerformQueryForRawQuery(findCtx, rawQuery, ".ds-"+otelIndex+"*", es)
-			assert.NoError(ct, err)
-			assert.GreaterOrEqual(ct, otelDocs.Hits.Total.Value, 1, "expected at least 1 otel document, got %d", otelDocs.Hits.Total.Value)
-
-			filebeatDocs, err = estools.PerformQueryForRawQuery(findCtx, rawQuery, ".ds-"+fbIndex+"*", es)
-			assert.NoError(ct, err)
-			assert.GreaterOrEqual(ct, filebeatDocs.Hits.Total.Value, 1, "expected at least 1 filebeat document, got %d", filebeatDocs.Hits.Total.Value)
-		},
-		3*time.Minute, 1*time.Second, "expected at least 1 document for both filebeat and otel modes")
-
-	filebeatDoc := filebeatDocs.Hits.Hits[0].Source
-	otelDoc := otelDocs.Hits.Hits[0].Source
-	ignoredFields := []string{
-		// Expected to change between agentDocs and OtelDocs
-		"@timestamp",
-		"agent.ephemeral_id",
-		"agent.id",
-		"event.created",
-	}
-
-	oteltest.AssertMapsEqual(t, filebeatDoc, otelDoc, ignoredFields, "expected documents to be equal")
-}
